requirements.txt
django
djangorestframework<3
